feat: restructure category DTOs and update service imports for consistency
