+++ src/core/blog/category/dto/index.ts
@@ -1,0 +1,1 @@
+export * from './category';
