+++ src/core/blog/category/services/update.service.ts
@@ -1,3 +1,4 @@
+import type { ICategoryDTO } from '@/core/blog/category/dto'
 import type { AppEither } from '@/core/error/app-either.protocols'
 import { isLeft, left, right } from '@/core/error/either'
 import {
@@ -8,12 +9,11 @@ import {
 } from '@/core/error/errors'
 import { db } from '@/infra/db'
 import { category } from '@/infra/db/schemas/blog'
+import { ensureAuthenticated } from '@/infra/helpers/auth'
 import { extractAndValidatePathParam } from '@/infra/helpers/params'
 import { categorySchema } from '@/infra/validations/schemas'
 import { and, eq, ne } from 'drizzle-orm'
 import { z } from 'zod'
-import type { ICategoryDTO } from '../dto/category.schema'
-import { ensureAuthenticated } from '@/infra/helpers/auth'
 
 const pathParamSchema = z.object({
   id: z.string().uuid('Invalid category ID')
