+++ src/core/blog/user/dto/index.ts
@@ -1,0 +1,2 @@
+export * from './role';
+export * from './user';
