+++ src/core/blog/user/dto/role.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+  USER = 'user',
+  EDITOR = 'editor',
+  ADMIn = 'admin'
+}
