+++ src/core/blog/user/dto/user.ts
@@ -1,0 +1,3 @@
+import type { DUser } from '@/infra/db/schemas/auth'
+
+export interface IUserDTO extends DUser {}
